refactor: remove dead code across mobile, web, and worker (#3)
Codebase cleanup pass 3 using knip analysis:

- apps/web: remove AnchorButton, QueryBoundary, SidebarLink, Field from
  components.tsx (unused components); remove wizardSteps, wizardStepLabels,
  nextWizardStep, previousWizardStep from onboarding.ts (unused navigation
  helpers)
- apps/mobile: remove unused constants (THIRTY_SECONDS_MS, ONE_SECOND_MS,
  DEFAULT_SYNC_COOLDOWN_SECONDS, MAX_OFFLINE_RETRIES, MAX_AUTH_RETRIES)
  from query.ts; remove withToast from toast-utils.ts
- apps/worker: remove cleanupOAuthState from oauth-state.ts; remove 7
  unused timestamp utility functions from timestamps.ts; remove
  optionalAuthMiddleware from middleware/auth.ts; trim polling/health.ts
  from 523 to 49 lines by removing 7 dead connection management functions
  (handlePollingAuthError, markConnectionExpired, markConnectionRevoked,
  resolveConnectionNotifications, trackPollFailure, clearPollFailures,
  resolvePollFailureNotifications) that were never wired to any caller

All typecheck, test, and lint checks pass.
