+++ apps/mobile/components/creator/CreatorBookmarks.tsx
@@ -13,8 +13,7 @@ import { Text } from '@/components/primitives';
 import { Radius, Typography, Spacing, type ThemeColors } from '@/constants/theme';
 import { useAppTheme } from '@/hooks/use-app-theme';
 import { useCreatorBookmarks } from '@/hooks/use-creator';
-import { mapContentType, mapProvider } from '@/hooks/use-items-trpc';
-import type { ContentType, Provider } from '@/lib/content-utils';
+import { mapContentType, mapProvider, type ContentType, type Provider } from '@/lib/content-utils';
 
 // ============================================================================
 // Types
