+++ apps/mobile/hooks/use-items-trpc.test.ts
@@ -63,7 +63,6 @@ import {
   useBookmarkItem,
   useArchiveItem,
   useToggleFinished,
-  formatDuration,
 } from './use-items-trpc';
 
 function createMockItem(overrides: Record<string, unknown> = {}) {
@@ -338,27 +337,6 @@ function createToggleUtils(initial?: {
     },
   };
 }
-
-describe('formatDuration', () => {
-  it('formats minute and hour durations', () => {
-    expect(formatDuration(45)).toBe('0:45');
-    expect(formatDuration(125)).toBe('2:05');
-    expect(formatDuration(3661)).toBe('1:01:01');
-  });
-
-  it('floors fractional seconds to avoid decimal output', () => {
-    expect(formatDuration(125.9)).toBe('2:05');
-    expect(formatDuration(59.999)).toBe('0:59');
-  });
-
-  it('returns undefined for invalid durations', () => {
-    expect(formatDuration(undefined)).toBeUndefined();
-    expect(formatDuration(null)).toBeUndefined();
-    expect(formatDuration(Number.NaN)).toBeUndefined();
-    expect(formatDuration(Number.POSITIVE_INFINITY)).toBeUndefined();
-    expect(formatDuration(-1)).toBeUndefined();
-  });
-});
 
 beforeEach(() => {
   jest.clearAllMocks();
