+++ apps/mobile/hooks/use-items-trpc.ts
@@ -8,22 +8,11 @@
 import { keepPreviousData } from '@tanstack/react-query';
 import { useRef } from 'react';
 import { trpc } from '@/lib/trpc';
-import {
-  ContentType,
-  Provider,
-  UserItemState,
-  formatDurationTimestamp,
-  mapContentType as mapSharedContentType,
-  mapProvider as mapSharedProvider,
-  type UIContentType,
-  type UIProvider,
-} from '@zine/shared';
+import { ContentType, Provider, UserItemState } from '@zine/shared';
 
 // ============================================================================
 // Types
 // ============================================================================
-
-export type { UIContentType, UIProvider };
 
 // ============================================================================
 // Optimistic Update Types
@@ -258,54 +247,6 @@ function createOptimisticConfig<TInput extends { id: string }>(
       invalidateRecapQueries(utils);
     },
   };
-}
-
-// ============================================================================
-// Helper Functions
-// ============================================================================
-
-/**
- * Map ContentType enum to UI-friendly lowercase string
- *
- * @param contentType - The ContentType enum value
- * @returns Lowercase string for UI display
- *
- * @example
- * mapContentType(ContentType.VIDEO) // => 'video'
- * mapContentType(ContentType.PODCAST) // => 'podcast'
- */
-export function mapContentType(contentType: ContentType): UIContentType {
-  return mapSharedContentType(contentType);
-}
-
-/**
- * Map Provider enum to UI-friendly lowercase string
- *
- * @param provider - The Provider enum value
- * @returns Lowercase string for UI display
- *
- * @example
- * mapProvider(Provider.YOUTUBE) // => 'youtube'
- * mapProvider(Provider.SPOTIFY) // => 'spotify'
- */
-export function mapProvider(provider: Provider): UIProvider {
-  return mapSharedProvider(provider);
-}
-
-/**
- * Format duration in seconds to human-readable string
- *
- * @param seconds - Duration in seconds (optional)
- * @returns Formatted duration string (H:MM:SS or M:SS) or undefined
- *
- * @example
- * formatDuration(3661) // => '1:01:01'
- * formatDuration(125) // => '2:05'
- * formatDuration(45) // => '0:45'
- * formatDuration(undefined) // => undefined
- */
-export function formatDuration(seconds?: number | null): string | undefined {
-  return formatDurationTimestamp(seconds);
 }
 
 function buildInboxItemsInput(options?: InboxItemsOptions) {
