+++ apps/mobile/lib/toast-utils.ts
@@ -133,45 +133,3 @@ export function showError(
     duration: ERROR_DURATION,
   });
 }
-
-// ============================================================================
-// Mutation Helpers
-// ============================================================================
-
-/**
- * Handle a mutation with automatic toast feedback.
- *
- * Shows success toast on completion, error toast on failure.
- *
- * @param toast - The toast manager from useToast()
- * @param mutation - The async mutation to execute
- * @param options - Configuration for success and error messages
- *
- * @example
- * ```tsx
- * await withToast(toast, () => toggleFinished(item.id), {
- *   successMessage: item.isFinished ? 'Marked as unfinished' : 'Marked as finished',
- *   errorMessage: 'Failed to update item',
- *   context: 'toggleFinished',
- * });
- * ```
- */
-export async function withToast<T>(
-  toast: ToastManager,
-  mutation: () => Promise<T>,
-  options: {
-    successMessage: string;
-    successDescription?: string;
-    errorMessage: string;
-    context?: string;
-  }
-): Promise<T | undefined> {
-  try {
-    const result = await mutation();
-    showSuccess(toast, options.successMessage, options.successDescription);
-    return result;
-  } catch (error) {
-    showError(toast, error, options.errorMessage, options.context);
-    return undefined;
-  }
-}
