+++ apps/mobile/lib/trpc-types.ts
@@ -27,3 +27,5 @@ export type RssListOutput = RouterOutputs['subscriptions']['rss']['list'];
 export type RssStatsOutput = RouterOutputs['subscriptions']['rss']['stats'];
 export type RssDiscoverInput = RouterInputs['subscriptions']['rss']['discover'];
 export type RssDiscoverOutput = RouterOutputs['subscriptions']['rss']['discover'];
+
+export type SyncStatusOutput = RouterOutputs['subscriptions']['syncStatus'];
