+++ apps/worker/src/lib/timestamps.ts
@@ -18,19 +18,6 @@
 // ============================================================================
 
 /**
- * Convert ISO 8601 string to Unix milliseconds.
- *
- * @param iso - ISO 8601 date string (e.g., "2024-01-15T10:00:00.000Z")
- * @returns Unix timestamp in milliseconds
- *
- * @example
- * isoToUnix("2024-01-15T10:00:00.000Z") // 1705312800000
- */
-export function isoToUnix(iso: string): number {
-  return new Date(iso).getTime();
-}
-
-/**
  * Convert Unix milliseconds to ISO 8601 string.
  *
  * @param unix - Unix timestamp in milliseconds
@@ -41,94 +28,6 @@ export function isoToUnix(iso: string): number {
  */
 export function unixToIso(unix: number): string {
   return new Date(unix).toISOString();
-}
-
-// ============================================================================
-// Current Timestamp Functions
-// ============================================================================
-
-/**
- * Get current timestamp as Unix milliseconds.
- *
- * Use this for new subscription-related tables.
- *
- * @returns Current Unix timestamp in milliseconds
- *
- * @example
- * const createdAt = nowUnix(); // 1705312800000
- */
-export function nowUnix(): number {
-  return Date.now();
-}
-
-/**
- * Get current timestamp as ISO 8601 string.
- *
- * Use this for legacy tables (users, items, user_items, sources).
- *
- * @returns Current ISO 8601 date string
- *
- * @example
- * const createdAt = nowIso(); // "2024-01-15T10:00:00.000Z"
- */
-export function nowIso(): string {
-  return new Date().toISOString();
-}
-
-// ============================================================================
-// Validation Functions
-// ============================================================================
-
-/**
- * Check if a value is a valid Unix timestamp in milliseconds.
- *
- * @param value - Value to check
- * @returns true if value is a valid Unix timestamp
- */
-export function isUnixTimestamp(value: unknown): value is number {
-  if (typeof value !== 'number') return false;
-  // Reasonable range: 2000-01-01 to 2100-01-01
-  const MIN_TIMESTAMP = 946684800000; // 2000-01-01
-  const MAX_TIMESTAMP = 4102444800000; // 2100-01-01
-  return value >= MIN_TIMESTAMP && value <= MAX_TIMESTAMP;
-}
-
-/**
- * Check if a value is a valid ISO 8601 date string.
- *
- * @param value - Value to check
- * @returns true if value is a valid ISO 8601 string
- */
-export function isIsoString(value: unknown): value is string {
-  if (typeof value !== 'string') return false;
-  const date = new Date(value);
-  return !isNaN(date.getTime()) && value.includes('T');
-}
-
-// ============================================================================
-// Optional/Nullable Helpers
-// ============================================================================
-
-/**
- * Convert optional ISO string to Unix, returning null if input is null/undefined.
- *
- * @param iso - Optional ISO 8601 date string
- * @returns Unix timestamp or null
- */
-export function isoToUnixOptional(iso: string | null | undefined): number | null {
-  if (iso == null) return null;
-  return isoToUnix(iso);
-}
-
-/**
- * Convert optional Unix timestamp to ISO, returning null if input is null/undefined.
- *
- * @param unix - Optional Unix timestamp
- * @returns ISO 8601 string or null
- */
-export function unixToIsoOptional(unix: number | null | undefined): string | null {
-  if (unix == null) return null;
-  return unixToIso(unix);
 }
 
 // ============================================================================
