+++ apps/worker/src/middleware/auth.ts
@@ -158,53 +158,3 @@ export function authMiddleware(): MiddlewareHandler<Env> {
     await next();
   };
 }
-
-/**
- * Optional authentication middleware
- *
- * Similar to authMiddleware but doesn't fail if no token is provided.
- * Sets userId to null if no valid token is present.
- * Useful for routes that work differently for authenticated vs anonymous users.
- *
- * @example
- * ```typescript
- * app.use('/api/public/*', optionalAuthMiddleware());
- *
- * app.get('/api/public/profile', (c) => {
- *   const userId = c.get('userId');
- *   if (userId) {
- *     // Return personalized content
- *   } else {
- *     // Return anonymous content
- *   }
- * });
- * ```
- */
-export function optionalAuthMiddleware(): MiddlewareHandler<Env> {
-  return async (c, next) => {
-    const authHeader = c.req.header('Authorization');
-
-    // No auth header - that's okay, continue without userId
-    if (!authHeader || !authHeader.startsWith('Bearer ')) {
-      c.set('userId', null);
-      await next();
-      return;
-    }
-
-    const token = authHeader.slice(7);
-
-    if (!token) {
-      c.set('userId', null);
-      await next();
-      return;
-    }
-
-    const jwksUrl = getClerkJwksUrl(c.env);
-    const result = await verifyClerkToken(token, jwksUrl);
-
-    // Set userId if valid, null otherwise
-    c.set('userId', result.success ? result.userId : null);
-
-    await next();
-  };
-}
