+++ apps/worker/src/polling/health.ts
@@ -1,95 +1,10 @@
 /**
- * Connection Health Monitoring and Recovery
+ * Connection Health Monitoring: Error Classification
  *
- * Handles authentication errors during polling with automatic status transitions
- * and user notifications. Implements the connection status flow:
- *
- * ACTIVE ──(refresh fails)──→ EXPIRED
- *    ↑                          │
- *    │                          │ (user revoked)
- *    │                          ↓
- *    └───(user reconnects)─── REVOKED
- *
- * Key responsibilities:
- * - Classify polling authentication errors
- * - Update connection/subscription status on auth failures
- * - Create user notifications with deduplication
- * - Auto-resolve notifications on reconnect
- * - Track poll failures with threshold-based notifications
+ * Classifies polling authentication errors to determine appropriate recovery actions.
  *
  * @see /features/subscriptions/backend-spec.md - Section 6.4: Connection Health & Recovery
  */
-
-import { and, eq, inArray, isNull } from 'drizzle-orm';
-import { ulid } from 'ulid';
-import { providerConnections, subscriptions, userNotifications } from '../db/schema';
-import { healthLogger } from '../lib/logger';
-import {
-  getValidAccessToken,
-  type ProviderConnection,
-  type TokenRefreshEnv,
-} from '../lib/token-refresh';
-import type { Database } from '../db';
-
-// NOTE: User notifications persist in D1 to support dedupe and resolution.
-
-// ============================================================================
-// Types
-// ============================================================================
-
-/**
- * Environment bindings required for health monitoring
- * Extends TokenRefreshEnv to ensure compatibility
- */
-export interface HealthMonitorEnv {
-  /** D1 database for persistent storage */
-  DB: D1Database;
-  /** KV namespace for poll failure tracking */
-  OAUTH_STATE_KV: KVNamespace;
-  /** AES-256 encryption key for token operations */
-  ENCRYPTION_KEY: string;
-  /** Google OAuth client ID (for YouTube) */
-  GOOGLE_CLIENT_ID: string;
-  /** Google OAuth client secret (for YouTube) - optional for PKCE flows */
-  GOOGLE_CLIENT_SECRET?: string;
-  /** Spotify OAuth client ID */
-  SPOTIFY_CLIENT_ID: string;
-  /** Spotify OAuth client secret */
-  SPOTIFY_CLIENT_SECRET: string;
-}
-
-/**
- * Notification parameters for creating user notifications
- */
-interface NotificationParams {
-  type: 'connection_expired' | 'connection_revoked' | 'poll_failures';
-  provider: string;
-  reason?: string;
-}
-
-// ============================================================================
-// Constants
-// ============================================================================
-
-/** Number of consecutive poll failures before notifying user */
-const FAILURE_THRESHOLD = 3;
-
-/** Notification messages by type */
-const NOTIFICATION_MESSAGES = {
-  connection_expired: (provider: string) =>
-    `Your ${provider} connection has expired. Please reconnect to continue receiving updates.`,
-  connection_revoked: (provider: string) =>
-    `Your ${provider} access was revoked. Please reconnect to continue receiving updates.`,
-  poll_failures: (provider: string) =>
-    `We're having trouble checking for new ${provider} content. This may resolve automatically.`,
-};
-
-/** Notification titles by type */
-const NOTIFICATION_TITLES = {
-  connection_expired: 'Connection Expired',
-  connection_revoked: 'Connection Revoked',
-  poll_failures: 'Sync Issue',
-};
 
 // ============================================================================
 // Error Classification
@@ -130,394 +45,4 @@ export function isAccessRevokedError(error: Error): boolean {
     msg.includes('access revoked') ||
     msg.includes('insufficient permissions')
   );
-}
-
-// ============================================================================
-// Main Error Handler
-// ============================================================================
-
-/**
- * Handle authentication errors during polling
- *
- * This function implements the core error recovery logic:
- * 1. If token expired, try to refresh it
- * 2. If refresh fails with invalid_grant, mark connection as EXPIRED
- * 3. If access was revoked (403), mark connection as REVOKED
- * 4. Update related subscriptions to DISCONNECTED
- * 5. Create user notification
- *
- * @param connection - The provider connection that encountered an error
- * @param error - The error that occurred during polling
- * @param env - Environment bindings
- * @param db - Database instance
- *
- * @example
- * ```typescript
- * try {
- *   await pollYouTubeChannel(connection, channelId);
- * } catch (error) {
- *   await handlePollingAuthError(connection, error as Error, env, db);
- * }
- * ```
- */
-export async function handlePollingAuthError(
-  connection: ProviderConnection,
-  error: Error,
-  env: HealthMonitorEnv,
-  db: Database
-): Promise<void> {
-  if (isTokenExpiredError(error)) {
-    // Try to refresh the token first
-    try {
-      await getValidAccessToken(connection, env as TokenRefreshEnv);
-      // Success! Token was refreshed, no status change needed
-      return;
-    } catch (refreshError) {
-      // Refresh failed - check if it's permanent
-      if (refreshError instanceof Error && isRefreshTokenInvalid(refreshError)) {
-        await markConnectionExpired(connection.id, 'refresh_token_invalid', db);
-      }
-      // For transient errors, don't change status yet - will retry next poll
-    }
-  } else if (isAccessRevokedError(error)) {
-    // User revoked access at the provider
-    await markConnectionRevoked(connection.id, db);
-  }
-}
-
-// ============================================================================
-// Connection Status Updates
-// ============================================================================
-
-/**
- * Mark a connection as expired and update related subscriptions
- *
- * Called when token refresh fails due to invalid refresh token.
- * Creates a user notification prompting reconnection.
- *
- * @param connectionId - ID of the connection to mark as expired
- * @param reason - Reason for expiration (for logging/notifications)
- * @param db - Database instance
- */
-export async function markConnectionExpired(
-  connectionId: string,
-  reason: string,
-  db: Database
-): Promise<void> {
-  // Get connection details for notification
-  const conn = await db.query.providerConnections.findFirst({
-    where: eq(providerConnections.id, connectionId),
-    columns: { userId: true, provider: true },
-  });
-
-  if (!conn) return;
-
-  // Update connection status
-  await db
-    .update(providerConnections)
-    .set({ status: 'EXPIRED' })
-    .where(eq(providerConnections.id, connectionId));
-
-  // Mark related subscriptions as disconnected
-  await db
-    .update(subscriptions)
-    .set({ status: 'DISCONNECTED', updatedAt: Date.now() })
-    .where(
-      and(
-        eq(subscriptions.userId, conn.userId),
-        eq(subscriptions.provider, conn.provider),
-        eq(subscriptions.status, 'ACTIVE')
-      )
-    );
-
-  // Create notification for user
-  await createUserNotification(
-    conn.userId,
-    {
-      type: 'connection_expired',
-      provider: conn.provider,
-      reason,
-    },
-    db
-  );
-
-  healthLogger.info('Connection marked as EXPIRED', {
-    connectionId,
-    userId: conn.userId,
-    provider: conn.provider,
-    reason,
-  });
-}
-
-/**
- * Mark a connection as revoked and update related subscriptions
- *
- * Called when provider returns 403/access revoked error.
- * Creates a user notification prompting reconnection.
- *
- * @param connectionId - ID of the connection to mark as revoked
- * @param db - Database instance
- */
-export async function markConnectionRevoked(connectionId: string, db: Database): Promise<void> {
-  // Get connection details for notification
-  const conn = await db.query.providerConnections.findFirst({
-    where: eq(providerConnections.id, connectionId),
-    columns: { userId: true, provider: true },
-  });
-
-  if (!conn) return;
-
-  // Update connection status
-  await db
-    .update(providerConnections)
-    .set({ status: 'REVOKED' })
-    .where(eq(providerConnections.id, connectionId));
-
-  // Mark related subscriptions as disconnected
-  await db
-    .update(subscriptions)
-    .set({ status: 'DISCONNECTED', updatedAt: Date.now() })
-    .where(
-      and(
-        eq(subscriptions.userId, conn.userId),
-        eq(subscriptions.provider, conn.provider),
-        eq(subscriptions.status, 'ACTIVE')
-      )
-    );
-
-  // Create notification for user
-  await createUserNotification(
-    conn.userId,
-    {
-      type: 'connection_revoked',
-      provider: conn.provider,
-    },
-    db
-  );
-
-  healthLogger.info('Connection marked as REVOKED', {
-    connectionId,
-    userId: conn.userId,
-    provider: conn.provider,
-  });
-}
-
-// ============================================================================
-// User Notifications
-// ============================================================================
-
-/**
- * Create a user notification with deduplication
- *
- * Prevents duplicate active notifications of the same type/provider combo.
- * If an active notification exists, it won't create a new one.
- *
- * @param userId - User ID to notify
- * @param params - Notification parameters (type, provider, reason)
- * @param db - Database instance
- */
-async function createUserNotification(
-  userId: string,
-  params: NotificationParams,
-  db: Database
-): Promise<void> {
-  const title = NOTIFICATION_TITLES[params.type];
-  const message = NOTIFICATION_MESSAGES[params.type](params.provider);
-
-  const existing = await db.query.userNotifications.findFirst({
-    where: and(
-      eq(userNotifications.userId, userId),
-      eq(userNotifications.type, params.type),
-      eq(userNotifications.provider, params.provider),
-      isNull(userNotifications.resolvedAt)
-    ),
-    columns: { id: true },
-  });
-
-  if (existing) {
-    healthLogger.info('NOTIFICATION deduped (active already exists)', {
-      userId,
-      type: params.type,
-      provider: params.provider,
-    });
-    return;
-  }
-
-  const now = Date.now();
-  const data = params.reason ? JSON.stringify({ reason: params.reason }) : null;
-
-  await db.insert(userNotifications).values({
-    id: ulid(),
-    userId,
-    type: params.type,
-    provider: params.provider,
-    title,
-    message,
-    data,
-    createdAt: now,
-  });
-
-  healthLogger.info('NOTIFICATION created', {
-    userId,
-    title,
-    message,
-    reason: params.reason,
-  });
-}
-
-/**
- * Auto-resolve connection notifications on successful reconnect
- *
- * Called when a user successfully reconnects a provider.
- * Resolves any outstanding connection_expired or connection_revoked notifications.
- *
- * @param userId - User ID who reconnected
- * @param provider - Provider that was reconnected
- * @param db - Database instance
- *
- * @example
- * ```typescript
- * // In OAuth callback after successful token exchange:
- * await resolveConnectionNotifications(userId, 'YOUTUBE', db);
- * ```
- */
-export async function resolveConnectionNotifications(
-  userId: string,
-  provider: string,
-  db: Database
-): Promise<void> {
-  const now = Date.now();
-
-  await db
-    .update(userNotifications)
-    .set({ resolvedAt: now })
-    .where(
-      and(
-        eq(userNotifications.userId, userId),
-        eq(userNotifications.provider, provider),
-        inArray(userNotifications.type, ['connection_expired', 'connection_revoked']),
-        isNull(userNotifications.resolvedAt)
-      )
-    );
-
-  healthLogger.info('Resolved connection notifications', { userId, provider });
-}
-
-// ============================================================================
-// Poll Failure Tracking
-// ============================================================================
-
-/**
- * Track consecutive poll failures for a subscription
- *
- * Increments failure count in KV storage. After reaching FAILURE_THRESHOLD (3),
- * creates a notification to inform the user of sync issues.
- *
- * Failures auto-expire after 24 hours in KV.
- *
- * @param subscriptionId - The subscription that failed to poll
- * @param error - The error that occurred
- * @param env - Environment bindings (needs KV namespace)
- * @param db - Database instance (for notifications)
- *
- * @example
- * ```typescript
- * try {
- *   await pollSubscription(sub);
- *   await clearPollFailures(sub.id, env); // Clear on success
- * } catch (error) {
- *   await trackPollFailure(sub.id, error as Error, env, db);
- * }
- * ```
- */
-export async function trackPollFailure(
-  subscriptionId: string,
-  error: Error,
-  env: HealthMonitorEnv,
-  db: Database
-): Promise<void> {
-  const key = `poll:failures:${subscriptionId}`;
-  const current = parseInt((await env.OAUTH_STATE_KV.get(key)) || '0', 10);
-  const newCount = current + 1;
-
-  // Store with 24-hour expiration (failures auto-reset)
-  await env.OAUTH_STATE_KV.put(key, String(newCount), { expirationTtl: 24 * 3600 });
-
-  healthLogger.warn('Poll failure', {
-    failureCount: newCount,
-    subscriptionId,
-    error: error.message,
-  });
-
-  // Notify user after threshold is reached
-  if (newCount === FAILURE_THRESHOLD) {
-    const sub = await db.query.subscriptions.findFirst({
-      where: eq(subscriptions.id, subscriptionId),
-      columns: { userId: true, provider: true },
-    });
-
-    if (sub) {
-      await createUserNotification(
-        sub.userId,
-        {
-          type: 'poll_failures',
-          provider: sub.provider,
-        },
-        db
-      );
-    }
-  }
-}
-
-/**
- * Clear poll failures for a subscription on successful poll
- *
- * Resets the failure counter. Should be called after each successful poll.
- *
- * @param subscriptionId - The subscription that polled successfully
- * @param env - Environment bindings (needs KV namespace)
- *
- * @example
- * ```typescript
- * const items = await pollSubscription(sub);
- * await clearPollFailures(sub.id, env);
- * ```
- */
-export async function clearPollFailures(
-  subscriptionId: string,
-  env: HealthMonitorEnv
-): Promise<void> {
-  await env.OAUTH_STATE_KV.delete(`poll:failures:${subscriptionId}`);
-}
-
-/**
- * Resolve poll failure notifications on successful poll
- *
- * Called after a successful poll to auto-resolve any outstanding
- * poll_failures notifications for the subscription.
- *
- * @param userId - User ID who owns the subscription
- * @param provider - Provider of the subscription
- * @param db - Database instance
- */
-export async function resolvePollFailureNotifications(
-  userId: string,
-  provider: string,
-  db: Database
-): Promise<void> {
-  const now = Date.now();
-
-  await db
-    .update(userNotifications)
-    .set({ resolvedAt: now })
-    .where(
-      and(
-        eq(userNotifications.userId, userId),
-        eq(userNotifications.provider, provider),
-        eq(userNotifications.type, 'poll_failures'),
-        isNull(userNotifications.resolvedAt)
-      )
-    );
-
-  healthLogger.info('Resolved poll failure notifications', { userId, provider });
 }
