+++ packages/shared/src/types/domain.ts
@@ -42,6 +42,7 @@ export type UserItemStateValue = `${UserItemState}`;
 export type SubscriptionStatusValue = `${SubscriptionStatus}`;
 export type ProviderConnectionStatusValue = `${ProviderConnectionStatus}`;
 export type OAuthProvider = Extract<ProviderValue, 'YOUTUBE' | 'SPOTIFY' | 'GMAIL'>;
+export type SubscriptionSource = OAuthProvider | 'RSS';
 export interface Item {
   id: string;
   contentType: ContentType;
